code quality enhancement
In this commit, I made the following additions and modifications:

1. Renamed cpp/modmesh/io to cpp/modmesh/inout to prevent
   unnecessary ambiguity.
2. Renamed the namespace IO to inout and removed the Gmsh namespace.
3. Renamed the struct ElementDef to GmshElementDef.
4. Moved non-template functions from gmsh.hpp to gmsh.cpp.
5. The Gmsh constructor now takes a string input as data, and the
   python wrapper also takes py::bytes as input, therefore file IO
   is now taken care of by python.
