+++ cpp/modmesh/inout/CMakeLists.txt
@@ -1,0 +1,28 @@
+
+cmake_minimum_required(VERSION 3.16)
+
+set(MODMESH_INOUT_HEADERS
+    ${CMAKE_CURRENT_SOURCE_DIR}/inout.hpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/gmsh.hpp
+    CACHE FILEPATH "" FORCE)
+
+set(MODMESH_INOUT_SOURCES
+    ${CMAKE_CURRENT_SOURCE_DIR}/gmsh.cpp
+    CACHE FILEPATH "" FORCE)
+
+set(MODMESH_INOUT_PYMODHEADERS
+    ${CMAKE_CURRENT_SOURCE_DIR}/pymod/inout_pymod.hpp
+    CACHE FILEPATH "" FORCE)
+
+set(MODMESH_INOUT_PYMODSOURCES
+    ${CMAKE_CURRENT_SOURCE_DIR}/pymod/inout_pymod.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/pymod/wrap_Gmsh.cpp
+    CACHE FILEPATH "" FORCE)
+
+set(MODMESH_INOUT_FILES
+    ${MODMESH_INOUT_HEADERS}
+    ${MODMESH_INOUT_SOURCES}
+    ${MODMESH_INOUT_PYMODHEADERS}
+    ${MODMESH_INOUT_PYMODSOURCES}
+    CACHE FILEPATH "" FORCE)
+# vim: set ff=unix fenc=utf8 nobomb et sw=4 ts=4 sts=4:
