+++ cpp/modmesh/inout/inout.hpp
@@ -1,4 +1,4 @@
 #pragma once
-#include <modmesh/io/gmsh.hpp>
+#include <modmesh/inout/gmsh.hpp>
 
 // vim: set ff=unix fenc=utf8 et sw=4 ts=4 sts=4:
