+++ cpp/modmesh/inout/pymod/wrap_Gmsh.cpp
@@ -1,4 +1,4 @@
-#include <modmesh/io/pymod/io_pymod.hpp>
+#include <modmesh/inout/pymod/inout_pymod.hpp>
 #include <modmesh/modmesh.hpp>
 
 namespace modmesh
@@ -8,11 +8,11 @@ namespace python
 {
 
 class MODMESH_PYTHON_WRAPPER_VISIBILITY WrapGmsh
-    : public WrapBase<WrapGmsh, IO::Gmsh::Gmsh, std::shared_ptr<IO::Gmsh::Gmsh>>
+    : public WrapBase<WrapGmsh, inout::Gmsh, std::shared_ptr<inout::Gmsh>>
 {
 public:
 
-    using base_type = WrapBase<WrapGmsh, IO::Gmsh::Gmsh, std::shared_ptr<IO::Gmsh::Gmsh>>;
+    using base_type = WrapBase<WrapGmsh, inout::Gmsh, std::shared_ptr<inout::Gmsh>>;
     using wrapped_type = typename base_type::wrapped_type;
 
     friend root_base_type;
@@ -20,16 +20,16 @@ public:
 protected:
 
     WrapGmsh(pybind11::module & mod, char const * pyname, char const * pydoc)
-        : WrapBase<WrapGmsh, IO::Gmsh::Gmsh, std::shared_ptr<IO::Gmsh::Gmsh>>(mod, pyname, pydoc)
+        : WrapBase<WrapGmsh, inout::Gmsh, std::shared_ptr<inout::Gmsh>>(mod, pyname, pydoc)
     {
         namespace py = pybind11; // NOLINT(misc-unused-alias-decls)
 
         (*this)
             .def(
                 py::init(
-                    [](const std::string & path)
-                    { return std::make_shared<IO::Gmsh::Gmsh>(path); }),
-                py::arg("file_name"))
+                    [](const py::bytes & data)
+                    { return std::make_shared<inout::Gmsh>(data); }),
+                py::arg("data"))
             .def("toblock", &wrapped_type::toblock);
         ;
     }
