--- cpp/modmesh/io/CMakeLists.txt
@@ -1,23 +1,0 @@
-
-cmake_minimum_required(VERSION 3.16)
-
-set(MODMESH_IO_HEADERS
-    ${CMAKE_CURRENT_SOURCE_DIR}/io.hpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/gmsh.hpp
-    CACHE FILEPATH "" FORCE)
-
-set(MODMESH_IO_PYMODHEADERS
-    ${CMAKE_CURRENT_SOURCE_DIR}/pymod/io_pymod.hpp
-    CACHE FILEPATH "" FORCE)
-
-set(MODMESH_IO_PYMODSOURCES
-    ${CMAKE_CURRENT_SOURCE_DIR}/pymod/io_pymod.cpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/pymod/wrap_Gmsh.cpp
-    CACHE FILEPATH "" FORCE)
-
-set(MODMESH_IO_FILES
-    ${MODMESH_IO_HEADERS}
-    ${MODMESH_IO_PYMODHEADERS}
-    ${MODMESH_IO_PYMODSOURCES}
-    CACHE FILEPATH "" FORCE)
-# vim: set ff=unix fenc=utf8 nobomb et sw=4 ts=4 sts=4:
