+++ gtests/test_nopython_inout.cpp
@@ -1,4 +1,4 @@
-#include <modmesh/io/io.hpp>
+#include <modmesh/inout/inout.hpp>
 
 #include <gtest/gtest.h>
 #include <gmock/gmock.h>
@@ -11,8 +11,7 @@ std::string g_test_file_path;
 
 TEST(Gmsh_Parser, NonCellTypeDefinition)
 {
-    namespace gmsh = modmesh::IO::Gmsh;
-    auto ele_def = gmsh::ElementDef::by_id(0);
+    auto ele_def = modmesh::inout::GmshElementDef::by_id(0);
     EXPECT_EQ(ele_def.ndim(), 0);
     EXPECT_EQ(ele_def.nnds(), 0);
     EXPECT_EQ(ele_def.mmtpn(), 0);
@@ -21,8 +20,7 @@ TEST(Gmsh_Parser, NonCellTypeDefinition)
 
 TEST(Gmsh_Parser, Line2NodeDefinition)
 {
-    namespace gmsh = modmesh::IO::Gmsh;
-    auto ele_def = gmsh::ElementDef::by_id(1);
+    auto ele_def = modmesh::inout::GmshElementDef::by_id(1);
     EXPECT_EQ(ele_def.ndim(), 1);
     EXPECT_EQ(ele_def.nnds(), 2);
     EXPECT_EQ(ele_def.mmtpn(), 2);
@@ -31,8 +29,7 @@ TEST(Gmsh_Parser, Line2NodeDefinition)
 
 TEST(Gmsh_Parser, Triangle3NodeDefinition)
 {
-    namespace gmsh = modmesh::IO::Gmsh;
-    auto ele_def = gmsh::ElementDef::by_id(2);
+    auto ele_def = modmesh::inout::GmshElementDef::by_id(2);
     EXPECT_EQ(ele_def.ndim(), 2);
     EXPECT_EQ(ele_def.nnds(), 3);
     EXPECT_EQ(ele_def.mmtpn(), 4);
@@ -41,8 +38,7 @@ TEST(Gmsh_Parser, Triangle3NodeDefinition)
 
 TEST(Gmsh_Parser, Quadrangle4NodeDefinition)
 {
-    namespace gmsh = modmesh::IO::Gmsh;
-    auto ele_def = gmsh::ElementDef::by_id(3);
+    auto ele_def = modmesh::inout::GmshElementDef::by_id(3);
     EXPECT_EQ(ele_def.ndim(), 2);
     EXPECT_EQ(ele_def.nnds(), 4);
     EXPECT_EQ(ele_def.mmtpn(), 3);
@@ -51,8 +47,7 @@ TEST(Gmsh_Parser, Quadrangle4NodeDefinition)
 
 TEST(Gmsh_Parser, Tetrahedron4NodeDefinition)
 {
-    namespace gmsh = modmesh::IO::Gmsh;
-    auto ele_def = gmsh::ElementDef::by_id(4);
+    auto ele_def = modmesh::inout::GmshElementDef::by_id(4);
     EXPECT_EQ(ele_def.ndim(), 3);
     EXPECT_EQ(ele_def.nnds(), 4);
     EXPECT_EQ(ele_def.mmtpn(), 6);
@@ -61,8 +56,7 @@ TEST(Gmsh_Parser, Tetrahedron4NodeDefinition)
 
 TEST(Gmsh_Parser, Hexahedron8NodeDefinition)
 {
-    namespace gmsh = modmesh::IO::Gmsh;
-    auto ele_def = gmsh::ElementDef::by_id(5);
+    auto ele_def = modmesh::inout::GmshElementDef::by_id(5);
     EXPECT_EQ(ele_def.ndim(), 3);
     EXPECT_EQ(ele_def.nnds(), 8);
     EXPECT_EQ(ele_def.mmtpn(), 5);
@@ -71,8 +65,7 @@ TEST(Gmsh_Parser, Hexahedron8NodeDefinition)
 
 TEST(Gmsh_Parser, Prism6NodeDefinition)
 {
-    namespace gmsh = modmesh::IO::Gmsh;
-    auto ele_def = gmsh::ElementDef::by_id(6);
+    auto ele_def = modmesh::inout::GmshElementDef::by_id(6);
     EXPECT_EQ(ele_def.ndim(), 3);
     EXPECT_EQ(ele_def.nnds(), 6);
     EXPECT_EQ(ele_def.mmtpn(), 7);
@@ -81,8 +74,7 @@ TEST(Gmsh_Parser, Prism6NodeDefinition)
 
 TEST(Gmsh_Parser, Pryamid5NodeDefinition)
 {
-    namespace gmsh = modmesh::IO::Gmsh;
-    auto ele_def = gmsh::ElementDef::by_id(7);
+    auto ele_def = modmesh::inout::GmshElementDef::by_id(7);
     EXPECT_EQ(ele_def.ndim(), 3);
     EXPECT_EQ(ele_def.nnds(), 5);
     EXPECT_EQ(ele_def.mmtpn(), 8);
@@ -91,8 +83,7 @@ TEST(Gmsh_Parser, Pryamid5NodeDefinition)
 
 TEST(Gmsh_Parser, Line3NodeDefinition)
 {
-    namespace gmsh = modmesh::IO::Gmsh;
-    auto ele_def = gmsh::ElementDef::by_id(8);
+    auto ele_def = modmesh::inout::GmshElementDef::by_id(8);
     EXPECT_EQ(ele_def.ndim(), 1);
     EXPECT_EQ(ele_def.nnds(), 3);
     EXPECT_EQ(ele_def.mmtpn(), 2);
@@ -101,8 +92,7 @@ TEST(Gmsh_Parser, Line3NodeDefinition)
 
 TEST(Gmsh_Parser, Triangle6NodeDefinition)
 {
-    namespace gmsh = modmesh::IO::Gmsh;
-    auto ele_def = gmsh::ElementDef::by_id(9);
+    auto ele_def = modmesh::inout::GmshElementDef::by_id(9);
     EXPECT_EQ(ele_def.ndim(), 2);
     EXPECT_EQ(ele_def.nnds(), 6);
     EXPECT_EQ(ele_def.mmtpn(), 4);
@@ -111,8 +101,7 @@ TEST(Gmsh_Parser, Triangle6NodeDefinition)
 
 TEST(Gmsh_Parser, Quadrangle9NodeDefinition)
 {
-    namespace gmsh = modmesh::IO::Gmsh;
-    auto ele_def = gmsh::ElementDef::by_id(10);
+    auto ele_def = modmesh::inout::GmshElementDef::by_id(10);
     EXPECT_EQ(ele_def.ndim(), 2);
     EXPECT_EQ(ele_def.nnds(), 9);
     EXPECT_EQ(ele_def.mmtpn(), 3);
@@ -121,8 +110,7 @@ TEST(Gmsh_Parser, Quadrangle9NodeDefinition)
 
 TEST(Gmsh_Parser, tetrahedron10NodeDefinition)
 {
-    namespace gmsh = modmesh::IO::Gmsh;
-    auto ele_def = gmsh::ElementDef::by_id(11);
+    auto ele_def = modmesh::inout::GmshElementDef::by_id(11);
     EXPECT_EQ(ele_def.ndim(), 3);
     EXPECT_EQ(ele_def.nnds(), 10);
     EXPECT_EQ(ele_def.mmtpn(), 6);
@@ -131,8 +119,7 @@ TEST(Gmsh_Parser, tetrahedron10NodeDefinition)
 
 TEST(Gmsh_Parser, hexahedron27NodeDefinition)
 {
-    namespace gmsh = modmesh::IO::Gmsh;
-    auto ele_def = gmsh::ElementDef::by_id(12);
+    auto ele_def = modmesh::inout::GmshElementDef::by_id(12);
     EXPECT_EQ(ele_def.ndim(), 3);
     EXPECT_EQ(ele_def.nnds(), 27);
     EXPECT_EQ(ele_def.mmtpn(), 5);
@@ -141,8 +128,7 @@ TEST(Gmsh_Parser, hexahedron27NodeDefinition)
 
 TEST(Gmsh_Parser, Prism18NodeDefinition)
 {
-    namespace gmsh = modmesh::IO::Gmsh;
-    auto ele_def = gmsh::ElementDef::by_id(13);
+    auto ele_def = modmesh::inout::GmshElementDef::by_id(13);
     EXPECT_EQ(ele_def.ndim(), 3);
     EXPECT_EQ(ele_def.nnds(), 18);
     EXPECT_EQ(ele_def.mmtpn(), 7);
@@ -151,8 +137,7 @@ TEST(Gmsh_Parser, Prism18NodeDefinition)
 
 TEST(Gmsh_Parser, Pyramid14NodeDefinition)
 {
-    namespace gmsh = modmesh::IO::Gmsh;
-    auto ele_def = gmsh::ElementDef::by_id(14);
+    auto ele_def = modmesh::inout::GmshElementDef::by_id(14);
     EXPECT_EQ(ele_def.ndim(), 3);
     EXPECT_EQ(ele_def.nnds(), 14);
     EXPECT_EQ(ele_def.mmtpn(), 8);
@@ -161,8 +146,7 @@ TEST(Gmsh_Parser, Pyramid14NodeDefinition)
 
 TEST(Gmsh_Parser, Point1NodeDefinition)
 {
-    namespace gmsh = modmesh::IO::Gmsh;
-    auto ele_def = gmsh::ElementDef::by_id(15);
+    auto ele_def = modmesh::inout::GmshElementDef::by_id(15);
     EXPECT_EQ(ele_def.ndim(), 0);
     EXPECT_EQ(ele_def.nnds(), 1);
     EXPECT_EQ(ele_def.mmtpn(), 1);
@@ -171,8 +155,7 @@ TEST(Gmsh_Parser, Point1NodeDefinition)
 
 TEST(Gmsh_Parser, Quadrangle8NodeDefinition)
 {
-    namespace gmsh = modmesh::IO::Gmsh;
-    auto ele_def = gmsh::ElementDef::by_id(16);
+    auto ele_def = modmesh::inout::GmshElementDef::by_id(16);
     EXPECT_EQ(ele_def.ndim(), 2);
     EXPECT_EQ(ele_def.nnds(), 8);
     EXPECT_EQ(ele_def.mmtpn(), 3);
@@ -181,8 +164,7 @@ TEST(Gmsh_Parser, Quadrangle8NodeDefinition)
 
 TEST(Gmsh_Parser, hexahedron20NodeDefinition)
 {
-    namespace gmsh = modmesh::IO::Gmsh;
-    auto ele_def = gmsh::ElementDef::by_id(17);
+    auto ele_def = modmesh::inout::GmshElementDef::by_id(17);
     EXPECT_EQ(ele_def.ndim(), 3);
     EXPECT_EQ(ele_def.nnds(), 20);
     EXPECT_EQ(ele_def.mmtpn(), 5);
@@ -191,8 +173,7 @@ TEST(Gmsh_Parser, hexahedron20NodeDefinition)
 
 TEST(Gmsh_Parser, Prism15NodeDefinition)
 {
-    namespace gmsh = modmesh::IO::Gmsh;
-    auto ele_def = gmsh::ElementDef::by_id(18);
+    auto ele_def = modmesh::inout::GmshElementDef::by_id(18);
     EXPECT_EQ(ele_def.ndim(), 3);
     EXPECT_EQ(ele_def.nnds(), 15);
     EXPECT_EQ(ele_def.mmtpn(), 7);
@@ -201,8 +182,7 @@ TEST(Gmsh_Parser, Prism15NodeDefinition)
 
 TEST(Gmsh_Parser, Pyramid13NodeDefinition)
 {
-    namespace gmsh = modmesh::IO::Gmsh;
-    auto ele_def = gmsh::ElementDef::by_id(19);
+    auto ele_def = modmesh::inout::GmshElementDef::by_id(19);
     EXPECT_EQ(ele_def.ndim(), 3);
     EXPECT_EQ(ele_def.nnds(), 13);
     EXPECT_EQ(ele_def.mmtpn(), 8);
@@ -211,8 +191,7 @@ TEST(Gmsh_Parser, Pyramid13NodeDefinition)
 
 TEST(Gmsh_Parser, IncompleteTriangle9NodeDefinition)
 {
-    namespace gmsh = modmesh::IO::Gmsh;
-    auto ele_def = gmsh::ElementDef::by_id(20);
+    auto ele_def = modmesh::inout::GmshElementDef::by_id(20);
     EXPECT_EQ(ele_def.ndim(), 2);
     EXPECT_EQ(ele_def.nnds(), 9);
     EXPECT_EQ(ele_def.mmtpn(), 4);
@@ -221,8 +200,7 @@ TEST(Gmsh_Parser, IncompleteTriangle9NodeDefinition)
 
 TEST(Gmsh_Parser, Triangle10NodeDefinition)
 {
-    namespace gmsh = modmesh::IO::Gmsh;
-    auto ele_def = gmsh::ElementDef::by_id(21);
+    auto ele_def = modmesh::inout::GmshElementDef::by_id(21);
     EXPECT_EQ(ele_def.ndim(), 2);
     EXPECT_EQ(ele_def.nnds(), 10);
     EXPECT_EQ(ele_def.mmtpn(), 4);
@@ -231,8 +209,7 @@ TEST(Gmsh_Parser, Triangle10NodeDefinition)
 
 TEST(Gmsh_Parser, IncompleteTriangle12NodeDefinition)
 {
-    namespace gmsh = modmesh::IO::Gmsh;
-    auto ele_def = gmsh::ElementDef::by_id(22);
+    auto ele_def = modmesh::inout::GmshElementDef::by_id(22);
     EXPECT_EQ(ele_def.ndim(), 2);
     EXPECT_EQ(ele_def.nnds(), 12);
     EXPECT_EQ(ele_def.mmtpn(), 4);
@@ -241,8 +218,7 @@ TEST(Gmsh_Parser, IncompleteTriangle12NodeDefinition)
 
 TEST(Gmsh_Parser, Triangle15NodeDefinition)
 {
-    namespace gmsh = modmesh::IO::Gmsh;
-    auto ele_def = gmsh::ElementDef::by_id(23);
+    auto ele_def = modmesh::inout::GmshElementDef::by_id(23);
     EXPECT_EQ(ele_def.ndim(), 2);
     EXPECT_EQ(ele_def.nnds(), 15);
     EXPECT_EQ(ele_def.mmtpn(), 4);
@@ -251,8 +227,7 @@ TEST(Gmsh_Parser, Triangle15NodeDefinition)
 
 TEST(Gmsh_Parser, IncompleteTriangle15NodeDefinition)
 {
-    namespace gmsh = modmesh::IO::Gmsh;
-    auto ele_def = gmsh::ElementDef::by_id(24);
+    auto ele_def = modmesh::inout::GmshElementDef::by_id(24);
     EXPECT_EQ(ele_def.ndim(), 2);
     EXPECT_EQ(ele_def.nnds(), 15);
     EXPECT_EQ(ele_def.mmtpn(), 4);
@@ -261,8 +236,7 @@ TEST(Gmsh_Parser, IncompleteTriangle15NodeDefinition)
 
 TEST(Gmsh_Parser, IncompleteTriangle21NodeDefinition)
 {
-    namespace gmsh = modmesh::IO::Gmsh;
-    auto ele_def = gmsh::ElementDef::by_id(25);
+    auto ele_def = modmesh::inout::GmshElementDef::by_id(25);
     EXPECT_EQ(ele_def.ndim(), 2);
     EXPECT_EQ(ele_def.nnds(), 21);
     EXPECT_EQ(ele_def.mmtpn(), 4);
@@ -271,8 +245,7 @@ TEST(Gmsh_Parser, IncompleteTriangle21NodeDefinition)
 
 TEST(Gmsh_Parser, Edge4NodeDefinition)
 {
-    namespace gmsh = modmesh::IO::Gmsh;
-    auto ele_def = gmsh::ElementDef::by_id(26);
+    auto ele_def = modmesh::inout::GmshElementDef::by_id(26);
     EXPECT_EQ(ele_def.ndim(), 1);
     EXPECT_EQ(ele_def.nnds(), 4);
     EXPECT_EQ(ele_def.mmtpn(), 2);
@@ -281,8 +254,7 @@ TEST(Gmsh_Parser, Edge4NodeDefinition)
 
 TEST(Gmsh_Parser, Edge5NodeDefinition)
 {
-    namespace gmsh = modmesh::IO::Gmsh;
-    auto ele_def = gmsh::ElementDef::by_id(27);
+    auto ele_def = modmesh::inout::GmshElementDef::by_id(27);
     EXPECT_EQ(ele_def.ndim(), 1);
     EXPECT_EQ(ele_def.nnds(), 5);
     EXPECT_EQ(ele_def.mmtpn(), 2);
@@ -291,8 +263,7 @@ TEST(Gmsh_Parser, Edge5NodeDefinition)
 
 TEST(Gmsh_Parser, Edge6NodeDefinition)
 {
-    namespace gmsh = modmesh::IO::Gmsh;
-    auto ele_def = gmsh::ElementDef::by_id(28);
+    auto ele_def = modmesh::inout::GmshElementDef::by_id(28);
     EXPECT_EQ(ele_def.ndim(), 1);
     EXPECT_EQ(ele_def.nnds(), 6);
     EXPECT_EQ(ele_def.mmtpn(), 2);
@@ -301,8 +272,7 @@ TEST(Gmsh_Parser, Edge6NodeDefinition)
 
 TEST(Gmsh_Parser, Tetrahedron20NodeDefinition)
 {
-    namespace gmsh = modmesh::IO::Gmsh;
-    auto ele_def = gmsh::ElementDef::by_id(29);
+    auto ele_def = modmesh::inout::GmshElementDef::by_id(29);
     EXPECT_EQ(ele_def.ndim(), 3);
     EXPECT_EQ(ele_def.nnds(), 20);
     EXPECT_EQ(ele_def.mmtpn(), 6);
@@ -311,8 +281,7 @@ TEST(Gmsh_Parser, Tetrahedron20NodeDefinition)
 
 TEST(Gmsh_Parser, Tetrahedron35NodeDefinition)
 {
-    namespace gmsh = modmesh::IO::Gmsh;
-    auto ele_def = gmsh::ElementDef::by_id(30);
+    auto ele_def = modmesh::inout::GmshElementDef::by_id(30);
     EXPECT_EQ(ele_def.ndim(), 3);
     EXPECT_EQ(ele_def.nnds(), 35);
     EXPECT_EQ(ele_def.mmtpn(), 6);
@@ -321,8 +290,7 @@ TEST(Gmsh_Parser, Tetrahedron35NodeDefinition)
 
 TEST(Gmsh_Parser, Tetrahedron56NodeDefinition)
 {
-    namespace gmsh = modmesh::IO::Gmsh;
-    auto ele_def = gmsh::ElementDef::by_id(31);
+    auto ele_def = modmesh::inout::GmshElementDef::by_id(31);
     EXPECT_EQ(ele_def.ndim(), 3);
     EXPECT_EQ(ele_def.nnds(), 56);
     EXPECT_EQ(ele_def.mmtpn(), 6);
@@ -331,8 +299,7 @@ TEST(Gmsh_Parser, Tetrahedron56NodeDefinition)
 
 TEST(Gmsh_Parser, Hexahedron64NodeDefinition)
 {
-    namespace gmsh = modmesh::IO::Gmsh;
-    auto ele_def = gmsh::ElementDef::by_id(92);
+    auto ele_def = modmesh::inout::GmshElementDef::by_id(92);
     EXPECT_EQ(ele_def.ndim(), 3);
     EXPECT_EQ(ele_def.nnds(), 64);
     EXPECT_EQ(ele_def.mmtpn(), 5);
@@ -341,8 +308,7 @@ TEST(Gmsh_Parser, Hexahedron64NodeDefinition)
 
 TEST(Gmsh_Parser, Hexahedron125NodeDefinition)
 {
-    namespace gmsh = modmesh::IO::Gmsh;
-    auto ele_def = gmsh::ElementDef::by_id(93);
+    auto ele_def = modmesh::inout::GmshElementDef::by_id(93);
     EXPECT_EQ(ele_def.ndim(), 3);
     EXPECT_EQ(ele_def.nnds(), 125);
     EXPECT_EQ(ele_def.mmtpn(), 5);
