+++ tests/test_gmsh.py
@@ -12,7 +12,10 @@ class GmshTC(unittest.TestCase):
     def test_gmsh_parsing(self):
         path = os.path.join(os.path.abspath(os.path.dirname(__file__)),
                             "data/gmsh_triangle.msh")
-        gmsh_instance = modmesh.core.Gmsh(path)
+
+        data = open(path, 'rb').read()
+
+        gmsh_instance = modmesh.core.Gmsh(data)
         blk = gmsh_instance.toblock()
 
         # Check nodes information
